use 'which' when accessing tools in the test suite
When running the Graphviz test suite on a installation that was built without
some optional components, calls to these missing components can incorrectly
resolve to those from a prior Graphviz installation that happens to be in your
`$PATH`. This issue was partly addressed in
515c86a923601db5cb704da93046800df7da030a. This change takes things further,
ensuring that invocations of such tools also land on the right version under
test.

Gitlab: #2201