cgraph: fix expansion watermark for node set
The node set was intended to expand at >70% occupancy. However, the LHS of this
expression was the division of two `size_t`s not `double`s. An invariant in this
code is `self->size <= self->capacity`. So the LHS was always 0 or 1. This had
the same effect as setting the occupancy watermark to 99.9999…%. The set would
only expand when completely full, resulting in pathological lookup performance
as the set filled.

Revisiting the workloads used for evaluation when this code was added in
4f59cbc484841cfe79b4116935799e5d72f8d103, neither are significantly affected by
this change. Presumably this is because neither involve graphs with more than
1024 * 0.7 = 716.8 nodes, the first time this watermark and subsequent expansion
should (and now does) apply.

Gitlab: #2475