fix arrowhead missing from tail-end of edge
In `make_flat_adj_edges`, when using ports, the graph is cloned. From
`cloneGraph`'s description:

"Create clone graph. It stores the global Agsyms, to be restored in
cleanupCloneGraph."

This process failed to store, set and restore the `E_dir` global,
causing the `id` of `Agsym_t` from the *original* graph to be
incorrectly used to look up it's value in the *cloned* graph's
dictionary. Instead of getting the `dir` attribute, this often got one
of the compass ports. Since Graphviz is very forgiving by nature, this
was interpreted as the default value of the `dir` attribute, i.e.,
`forward` instead of the actual value.

GitLab: fixes #2437
Reported-by: Steve Roush