CMake: fix: install a top level Graphviz configuration file
This appears to be¹ the way to solve both warts from the earlier commit that
added the initial support. This has been left separate rather than squashing
these two commits into one to “show my work.” I am not sure this is completely
correct, so demonstrating the path by which I arrived at this solution may be
useful for someone aiming to improve on this in future.

¹ https://stackoverflow.com/questions/66416594/is-it-possible-for-multiple-projects-to-export-targets-under-the-same-package-na

Gitlab: closes #2598
Reported-by: Peter Steneteg