CMake: fix pathplan’s include directories
This was fixed for gvc in dcc5ed93775b66339c14f297e35c91f1cf3dd43a, but the fix
was never propagated to pathplan. Apart from making pathplan usable when
imported as a subproject, this will avoid references to the source directory
ending up in export files in upcoming changes.

Gitlab: #1477, #2109, #2598