add an extra test case for CMake external usage
This demonstrates a quirk I have not yet solved. When an application links
against cgraph it wants to simply declare this alone, but cgraph depends on cdt.
Just declaring a cgraph dependency on cdt does not work because the exported cdt
target is not in scope when the user has only executed
`find_package(graphviz::cgraph)`. This will be fixed in an upcoming commit.