CMake: fix: strengthen bindings man page dependencies
When building in parallel, compilation would sometimes fail complaining that
various (generated) bindings man pages did not exist. This is a little puzzling
as the dependencies appeared correct. Consulting CMake docs¹ and discussion²
suggests a dependency on _both_ the generated file and its target are necessary
to make this work reliably.

¹ https://cmake.org/cmake/help/latest/command/add_custom_command.html#example-generating-files-for-multiple-targets
² https://discourse.cmake.org/t/add-custom-command-with-target-dependency-doesnt-rebuild/7029