API BREAK: cgraph: extra 'g_seq2' into its own field
41c218e33bfd9242625826b92293dd4f149eee22 introduced a hidden member of
`Agraph_t` that could be located through some pointer addition on `g_seq`. This
was a cute trick to avoid an API break but involved code that was not very
clear. With the upcoming release due to be an API break, we can take the
opportunity to unwind this technical debt into something more obvious.

`void *` is used rather than a stronger type to avoid having to define
`Agraphs_t` in a public header. Users are not expected to access this field
directly.

Gitlab: #2593