cgraph: fix: incorporate 'is_html' into reference-counted string equality
When operating on reference-counted strings internally, whether a string was a
regular string or an HTML-like string would not be taken into account. Mostly
this was invisible to users. However one visible side effect was that it was not
possible to create a regular string and an HTML-like string within the same
graph. Whichever was created first would be considered authoritative and the
second would be de-duped into the first.

This change discriminates between the two, leading to less surprising outcomes.

Part of this change involves incorporating the `is_html` value into the string’s
hash. This is not strictly necessary (we could just allow a regular string and
an HTML-like string with the same textual content to hash to the same value, and
then discover their inequality while scanning for a match), but it is slightly
more intuitive to include this. The way it is incorporated into the Murmur Hash
code is intended to roughly correspond to an extra contiguous byte following the
string data.

`refstrbind`’s call into `strdict_find` hard codes that it is seeking a non-HTML
string because it has no further information. This is a user-facing API that may
be passed arbitrary strings, so calling `aghtmlstr` is not safe. In practice, it
seems only used for IDs, which are non-HTML strings.

Gitlab: #2634, fixes #2089