cgraph write_canonstr: take an extra parameter to discriminate known 'refstr_t's
`write_canonstr` allocated its input `str` as a reference-counted string because
it was written to be agnostic as to whether the caller passed it a string that
originated from a `refstr_t` or not. It needs `s` to be a reference-counted
string in order to later check whether it is an HTML-like string.

This worked, but hides a latent problem that would have become a bug after
upcoming changes. `str` is allocated as a non-HTML string (because
`write_canonstr` had no way of determining whether to call `agstrdup` vs
`agstrdup_html`). But `str` is _already_ an HTML-like string, this logic will
allocate a new non-HTML string after we begin discriminating these two types of
strings in an upcoming commit. The present change pre-emptively fixes this
latent issue.

A side effect of this change is that `write_canonstr` becomes slightly faster by
avoiding bumping reference counts for known `refstr_t`-originating strings. No
attempt was made to be thorough here; the only calls that were updated to pass
`true` were the ones that may be dealing with HTML-like strings. It is possible
there is further optimization to be had by passing `true` in more calls to
`write_canonstr`.

Gitlab: #2089