cgraph: remove 'refstr_t.store'
This structure was defined in a way where `s` and `store` duplicated
information. `s` was always expected to point to `store`. This was not a mistake
but rather a constraint imposed by libcdt’s API. When using a null-terminated
string as a dictionary key, libcdt requires the key offset to reference a
`char *`. Thus an actual `char *` was required in this structure, not just a
`char[]`. Now that these are no longer stored in a libcdt dictionary, this
constraint is no longer present.

Since `refstr_t` was implemented, Graphviz has moved to being compiled with C99.
C99 offers so-called “flexible array members”¹ that give a more concise,
intuitive, and safer way of defining the variable-length array `s`.

This change saves `sizeof(char *)` bytes per `refstr_t` allocation.

¹ https://en.wikipedia.org/wiki/Flexible_array_member