API BREAK: cgraph: replace refstr’s use of libcdt with an inline implementation
When creating a regular string and an HTML-like string, there is no way to tell
them apart in the reference-counted string dictionary. The reason for this is
that the comparison function used for dictionary entries is `strcmp` on the
string data; i.e. it does not consider the `is_html` bit. Fixing this bug
within the existing libcdt-based dictionary is not possible. The reason for this
is that the libcdt API offers consumers one of the following:
  1. Entries with a key of a fixed size; or
  2. Entries with a key that is a null-terminated `char *`.
What we need is a third thing: a variable length (though still null-terminated)
part of the key plus an extra byte. It would be possible to adapt the libcdt API
to support this. But the changes would be so invasive, it seems simpler to
implement a new specialized data structure.

This change does not solve the above described bug, but takes a step towards it.

The dictionary implementation added in this commit is based on the set
implementation in lib/cgraph/node.c. These could be unified in future.

The version of Murmur Hash added in this commit was copied from public domain
sources.¹ For simplicity for future readers, the code comment references the
original implementation rather than this third-party intermediary.

¹ https://github.com/Smattr/rumur/blob/db735fc1f2a5f4032a21a825e6d886e44c97df91/rumur/resources/header.c#L502-L552

Gitlab: #2089