API BREAK: cgraph afstrfree: take an extra parameter 'is_html'
This is another step towards discriminating non-HTML and HTML-like strings. Once
these are separated, we cannot reliably free a string without knowing which of
these types it is.

This change involved surveying all existing `agstrfree` calls and trying to
reverse engineer which are freeing non-HTML strings and which are freeing
HTML-like strings.

Gitlab: #2089