show more information when test of layout in empty subgraph fails
During construction of some recent changes, a state was reached where the only
test failure was this test case on the ASan builds. The way this test and its
assertions were written resulted in the actual ASan message explaining the
failure was swallowed. This change should hopefully make such situations more
debuggable in future.

Gitlab: #2078, !4101