util: teach generic list how to interact with ASan
It was possible for out-of-bounds accesses within generic lists to go undetected
by Address Sanitizer because, from the perspective of the sanitizer, the entire
capacity of a generic list was allocated. E.g.

  ints_list_t xs = {0};          // xs.base == NULL
  ints_list_append(&xs, 42);     // xs.base == {42}
  (void)ints_list_pop_back(&xs); // xs.base == {××} (conceptually)
  int y = xs.base[0];            // access to conceptually deallocated ××

The access in the final line read the allocated but unpopulated first entry of
the backing array. ASan could not detect this because it believed `xs.base[0]`
to be valid, allocated memory.

This change causes the generic list’s API to inform ASan whenever it
conceptually allocates or deallocates. Not only does this catch situations like
the above, this also allows ASan to catch bugs within the generic list
implementation itself.