neatogen rk_random: squash MSVC negation warnings
When building this code, MSVC says:

  lib\neatogen\randomkit.c(106,57): warning C4146: unary minus operator applied to unsigned type, result still unsigned
  lib\neatogen\randomkit.c(110,61): warning C4146: unary minus operator applied to unsigned type, result still unsigned
  lib\neatogen\randomkit.c(113,61): warning C4146: unary minus operator applied to unsigned type, result still unsigned

Unary negation of an unsigned value is well-defined, but this warning is
intended to catch code that looks like accidental misuse. In this case it is
intentional, so we rephrase it into something intended to be equivalent.

This squashes the only remaining C4146 warnings.

Reported-by: MSVC 19.40.33811.0