remove overly prescriptive conditions for an ASan-based test to pass
It is unclear to me why 098e82b7893bf637f8b7458a597b4a290920110b added this
check. While it is true that the init_rank problems persist, that is not the bug
that was reported in the original issue. Rewinding indicates the original crash
is no longer reproducible on bd97cff688f7a7b85b6f1262e14eb1cac0862fcd, so
something appears to have changed in the surrounding ecosystem.

Gitlab: !3628, closes #1328
Reported-by: Google Autofuzz project