core plugin: inline referenced images when rendering `-Tsvg_inline`
When using the `-Tsvg_inline` renderer, it is implied that the output content is
destined for embedding in a web page. Contemporary browsers implement the
so-called “same-origin policy,” one side-effect of which is blocking transitive
content loads. Thus they block the loading of images referenced via a URL in
`<image xlink:href="…"/>` in an SVG. The present change works around this
limitation by embedding the content within the output data itself.

This does not perform transitive embedding. That is, if you use image A.svg that
references image B.svg, the generated SVG output will only embed A.svg. The
embedded data will still contain references to the external B.svg which will
fail to load in a web browser.

Gitlab: #2636