fix: when including an SVG, do not let 'viewBox' override 'height' and 'width'
SVG images support the attributes `height`¹ and `width`² for describing
dimensions of the image. Additionally the `viewBox`³ attribute can be used to
scale the SVG content. To me, the SVG spec implies the `viewBox` attribute is
not intended to influence overall image dimensions. Yet it seems some SVG tools
generate images without `height` and `width` but with `viewBox` and intend the
latter to imply image dimensions (see tests/EDA.svg for an example of this).

1d0a78c15dd2954871d53b21b4a8b5409857fa87 seems to have observed this and
adjusted Graphviz to interpret `viewBox`. However the logic terminated after
encountering `viewBox`, assuming this was authoritative. This is not the case
when the image also has `height` and/or `width`. This change is intended to fix
this discrepancy.

This change has no effect on either of the remaining problems:
  1. `viewBox` is incorrectly parsed. The last two components are treated as
     corner coordinates when they are actually width and height.
  2. The docs⁴ claim `height` and `width` are required, which is inconsistent
     with how the code currently behaves.

¹ https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/height
² https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/width
³ https://developer.mozilla.org/en-US/docs/Web/SVG/Attribute/viewBox
⁴ https://graphviz.org/docs/attrs/image/

Gitlab: #2352, #2636
Mantis: #2135
Reported-by: Mark Nahabedian