expr: remove 'swstate' fallback global
When encountering a `switch` statement, if allocating memory for the switch
failed, the global static memory backing `swstate` would be used. It is unclear
what value this added. If the `sw` allocation fails, it is almost certain the
follow up `sw->base` allocation will fail, leaving you with a degenerate
`Switch_t` instance. In theory you could successfully parse a switch containing
0 cases in this scenario, but it seems extremely unlikely user programs would
actually contain such things. The value of limping along in this way only to
eventually fail anyway seems non-existent.

Additionally relying on a global like this was a barrier to any future work on
multithreading.

Gitlab: #2558