extract optional 'double' type into utility library
The use of macros was dropped as it does not seem worth it until we have more
than one optional type implemented.

The type `optional_double_t` is not space-optimized. If it becomes relevant to
create a compact array of these, a scheme like NaN boxing¹ could be used.

¹ https://piotrduperas.com/posts/nan-boxing

Gitlab: https://gitlab.com/graphviz/graphviz/-/merge_requests/4209#note_2333071524
Suggested-by: Magnus Jacobsson <Magnus.Jacobsson@berotec.se>