Revert "cgraph agsubnode: inline and specialize 'agroot(g)' call"
This reverts commit 91049d316640e322f4409a4e8e84331c8703ae18. This change was a
slight degradation in readability and maintainability for the goal of
performance acceleration. Following this,
1ed67af7af325216b2d4422ff5f38d3aa058487f dramatically improved performance in a
way that subsumed this. Reverting this change has little effect on the runtime
of the #2593 test case (2.17s → 2.13s, though this apparent speed up is likely
within the noise), so we choose to regain the former readability.

Gitlab: #2593