gvusershape: fix parsing of JPEG file
The parsing incorrectly started 20 bytes into the file, because
`imagetype` had already read those bytes. This position was almost
always in the middle of a segment's payload data. This often led to
the wrong or no image size being detected.

Before this fix, parsing of the "2619_1_2.jpg" image, which is used in
`test_2619_1` when the `images` parameter is `2619_1`, seemingly
succeeded¹. This subsequently triggered a bug² when calling
`gdImageCreateFromJpeg` in Windows Debug builds. The reason for this
bug is currently unknown and remains even after this fix. Now that
Graphviz can parse all JPEG images, this remaining bug is triggered also
when the `images` parameter is `2619_2`, hence the simplified, but
broader, xfail. For the same reason, `test_2619` now fails with Windows
Debug builds.

¹ "Seemingly", since the determined size was bogus.
² This doesn't seem to be a Graphviz bug since this simple test
program, which is orthogonal to Graphviz, also fails³ in Windows Debug
builds:

```
int main(void) {
  FILE *f = fopen("tests/2619_1_2.jpg", "rb");

  gdImagePtr p = gdImageCreateFromJpeg(f);

  return p == NULL ? 1 : 0;
}
```

³ The failure is a crash with exit code 3221226505 when calling
gdImageCreateFromJpeg. 3221226505 = 0xC0000409 and means
STATUS_STACK_BUFFER_OVERRUN, although that might not be what's
actually happening, see
https://devblogs.microsoft.com/oldnewthing/20230731-00/?p=108505

GitLab: #2619.