tests: add a test case for failed parsing of JPEG images
Without the xfail, filtering the output of the test with
`grep -E "^E.*Assertion|_ test_"` shows:

```
____________________________ test_2619_3[2619.jpg] _____________________________
E       AssertionError: could not find an 'image' element in the SVG
__________________________ test_2619_3[2619_1_1.jpg] ___________________________
E       AssertionError: could not find an 'image' element in the SVG
__________________________ test_2619_3[2619_1_2.jpg] ___________________________
E       AssertionError: assert '14988px' == '3264px'
__________________________ test_2619_3[2619_1_3.jpg] ___________________________
E       AssertionError: could not find an 'image' element in the SVG
__________________________ test_2619_3[2619_2_1.jpg] ___________________________
E       AssertionError: could not find an 'image' element in the SVG
__________________________ test_2619_3[2619_2_2.jpg] ___________________________
E       AssertionError: could not find an 'image' element in the SVG
__________________________ test_2619_3[2619_2_3.jpg] ___________________________
E       AssertionError: could not find an 'image' element in the SVG
```

I.e., Graphviz detects it cannot parse any image¹ except
"2619_1_2.jpg", but the parsed size of that image is bogus. These
problems will be fixed in an upcoming commit.

¹ This causes the size of the JPEG image to be set to zero and
subsequently no image to be generated in the output.

GitLab: #2619