ortho: fix: remove assumptions about total sides of 'gcells'
The total number of sides of `gcells` (cells corresponding to graph nodes) was
assumed to be the same as the number of nodes in the graph. I do not know why
this would be guaranteed true, but I do not have a deep understanding of this
code. The test case for #2643 proves this assumption is false. Running it under
Address Sanitizer reports:

  ==90489==ERROR: AddressSanitizer: heap-buffer-overflow on address 0x611000030980 at pc 0x7ffff6c97ea4 bp 0x7fffffffe040 sp 0x7fffffffe030
  WRITE of size 8 at 0x611000030980 thread T0
      #0 0x7ffff6c97ea3  (/tmp/tmp.QatYeZgkso/lib+0x697ea3)
        [ $ addr2line -e /tmp/tmp.QatYeZgkso/lib/libgvc.so 0x697ea3
          lib/ortho/maze.c:397 ]
      #1 0x7ffff6c9ad2b  (/tmp/tmp.QatYeZgkso/lib+0x69ad2b)
        [ $ addr2line -e /tmp/tmp.QatYeZgkso/lib/libgvc.so 0x69ad2b
          lib/ortho/maze.c:500 ]
      #2 0x7ffff6cb0f22  (/tmp/tmp.QatYeZgkso/lib+0x6b0f22)
        [ $ addr2line -e /tmp/tmp.QatYeZgkso/lib/libgvc.so 0x6b0f22
          lib/ortho/ortho.c:1283 ]
      #3 0x7ffff1326e68 in spline_edges_ lib/neatogen/neatosplines.c:650
      #4 0x7ffff1328bb7 in splineEdges lib/neatogen/neatosplines.c:770
      #5 0x7ffff1328cad in spline_edges1 lib/neatogen/neatosplines.c:783
      #6 0x7ffff1328df5 in spline_edges0 lib/neatogen/neatosplines.c:812
      #7 0x7ffff132a162 in spline_edges lib/neatogen/neatosplines.c:848
      #8 0x7ffff12c65eb in twopi_layout lib/twopigen/twopiinit.c:134
      #9 0x7ffff6a35c84  (/tmp/tmp.QatYeZgkso/lib+0x435c84)
      #10 0x5555555555df in main cmd/dot/dot.c:72
      #11 0x7ffff5a29d8f in __libc_start_call_main ../sysdeps/nptl/libc_start_call_main.h:58
      #12 0x7ffff5a29e3f in __libc_start_main_impl ../csu/libc-start.c:392
      #13 0x555555555324 in _start (/tmp/tmp.QatYeZgkso/bin/dot+0x1324)

The lines missing location information due to #2526 have been supplemented with
information from `addr2line`.

This change removes this assumption and replaces this monolithic allocation with
individual allocations per-cell. These are managed as dynamic lists, expanding
as necessary for the number of encountered sides.

Fixing this problem uncovers a further crash with the #2643 test case, hence it
is not un-xfailed yet.

Gitlab: #2526, #2643
Reported-by: noneofyourbusiness