CI: standardize on '$ID' for operating system detection
Detection of the operating system in use relied on `$OSTYPE`, most probably
because this is a widely recommended technique on the internet. Recently MSYS2
has begun reporting itself as “cywgin”¹ that thwarts our attempts to distinguish
Cygwin vs MinGW. This change standardizes instead on `$ID` that seems to be
universally available and still allows distinguishing Cygwin vs MinGW. This ends
up being a little more verbose for Linux handling, but an improvement overall.

The `[[ … == bar* ]]` syntax is the way to check the prefix of a string in
Bash.² This also standardizes on bracing these expansions which was done
inconsistently previously.

Other CI scripts are left alone, still using other variables like `$ID_LIKE`.
Perhaps these should also be adjusted this way in future.

¹ https://www.msys2.org/news/#2025-02-14-moving-msys2-closer-to-cygwin
² https://stackoverflow.com/questions/2172352/in-bash-how-can-i-check-if-a-string-begins-with-some-value
  https://tldp.org/LDP/abs/html/comparison-ops.html

Gitlab: fixes #2642