common dfs_range: rewrite to avoid recursion
In graphs with deep chains of nodes, the recursion in `dfs_range` was exceeding
platform stack limits, causing Graphviz to crash. The intent of this change is
to rephrase `dfs_range` such that it is non-recursive, while preserving its
behavior.

This change is a minor degradation in runtime performance on examples like
test_MR_2854 (8m58s → 9m31s (~6% slow down). This seems worth it in order to be
able to process previously impossible inputs.

test_2646 now passes but takes a long time to do so, hence marking it “slow”.

----

Note that this change (recursively) preserves the optimization from
4b736d297bb1599451e89c2fde911d966a1db3cf. An obvious further improvement seems
to be removing the unnecessary loop restarts when the top of stack has not been
modified. I.e.:

  diff --git lib/common/ns.c lib/common/ns.c
  index a8c256186..888ecbd7b 100644
  --- lib/common/ns.c
  +++ lib/common/ns.c
  @@ -1208,9 +1208,9 @@ static int dfs_range(node_t * v, edge_t * par, int low)
                      ND_low(n) = s->lim;
                      const dfs_state_t next = {.v = n, .par = e, .lim = s->lim};
                      dfs_stack_push_back(&todo, next);
  +		    processed_child = true;
  +		    break;
                  }
  -		processed_child = true;
  -		break;
              }
          }
          if (processed_child) {
  @@ -1229,9 +1229,9 @@ static int dfs_range(node_t * v, edge_t * par, int low)
                      ND_low(n) = s->lim;
                      const dfs_state_t next = {.v = n, .par = e, .lim = s->lim};
                      dfs_stack_push_back(&todo, next);
  +		    processed_child = true;
  +		    break;
                  }
  -		processed_child = true;
  -		break;
              }
          }
          if (processed_child) {

However profiling reveals this is actually a slow down. It seems making the loop
restart conditional like this introduces a branch that is frequently
mispredicted, slowing overall execution.

Gitlab: fixes #2646
Reported-by: forum user tolliob