add a performance test case
The graph in this commit is taken from !2854 that went on to be merged as !2857.
Upcoming changes may affect the runtime of `dfs_range`/`dfs_range_init` and we
would like to ensure performance does not regress. An unoptimized Graphviz takes
~10m to process this example on Linux, so hopefully a 20m window should be
sufficient to validate its performance. Any future slow down incurred is
expected to be algorithmic; that is, its effect would be >2× and thus beyond
this 20m window.

Gitlab: !2854, !2857, #2646