common dfs_range_init: rewrite to avoid recursion
In graphs with deep chains of nodes, the recursion in `dfs_range_init` was
exceeding platform stack limits, causing Graphviz to crash. The intent of this
change is to rephrase `dfs_range_init` such that it is non-recursive, while
preserving its behavior.

This change does not have a significant effect on performance of representative
workloads like test_MR_2854 (9m03s → 8m58s (~1% speed up) in an optimized build;
likely within the noise).

This does not un-xfail test_2646 because addressing this issue uncovers a
further stack overflow in this workload.

Gitlab: #2646
Reported-by: forum user tolliob