add another performance test case
The #2095 test case graph was used to evaluate the performance gain in !2857. In
future, we would like to ensure the performance of processing this graph does
not regress. The graph is processed fast enough, we are unlikely to catch
regressions using this test. But it serves as a useful backstop against massive
performance degradation.

Gitlab: #2095, !2854, !2857, #2646