gvgen: seed RNG once and allow control from the command line
The way in which the random number generator (RNG) was being seeded had a number
of weaknesses:
  1. The seed was based on the current time. This meant it was impractical to
     regenerate a graph in the same way twice.
  2. The RNG was seeded twice when generating a rooted tree (`-R`), once prior
     to calling `makeRandomTree` and once when calling `makeRandom`.

This change addresses both of these issues, with fallback to seeding based on
the current time as before. The point in the program at which `srand` is called
has changed, but it is assumed no user can have been relying on this; to seed
the RNG identically within two different runs would previously have required
running from the exact same starting time and having the exact same scheduling
from the operating system, processor, etc.

Gitlab: #2640