CI: remove 32-bit MinGW jobs
The mingw-w64-i686-mesa package has become uninstallable on 32-bit MSYS2:

  + pacman -S --noconfirm --needed … mingw-w64-i686-mesa
  error: target not found: mingw-w64-i686-mesa

The cause appears to be that the 24.3.3-1 → 25.0.0-1 update to this package¹
removed support for 32-bit x86. The deprecation of 32-bit MSYS2 was apparently
announced on 2023-12-10² but not in a forum we were monitoring, so this removal
caught us by surprise.³ It seems no longer feasible to offer Graphviz on this
platform.

¹ https://github.com/msys2/MINGW-packages/commit/0e212be98f702308799f8918f666fc762e8bd167
² https://github.com/msys2/MINGW-packages/discussions/19326
³ https://github.com/msys2/MINGW-packages/issues/23516

Gitlab: closes #2649