compile_c, run_c: anticipate absolute paths to dynamic libraries to link
Upcoming changes will need to compile code that links directly against plugins
that are not in the linker’s search path. This change prepares the way for this.

Note that the `pkg-config` case incurs some extra complexity to try to preserve
link order. The upcoming change referred to above passes something like
`link=["cgraph", "gvc", "/path/to/libgvplugin_core.so",
"/path/to/libgvplugin_dot_layout.so"]`. The naïve way to deal with this would
have been:
  1. Accumulate all non-absolute-paths, then make one `pkg-config` call; or
  2. Make a `pkg-config` call each time a non-absolute-path is encountered.
Both of these have weaknesses:
  1. The `ldflags` added from cgraph and gvc are now placed _after_ the plugin
     paths.
  2. The `cflags` added from cgraph and gvc are now duplicated.
Neither are fatal problems, and the implementation used here is not perfect in
preserving ordering either, but it seems to reasonably balance complexity with
getting close to the caller’s intuition/intent.

Gitlab: #2648