ortho: remove unused 0-th segment index in permutations
Contrary to its description, `generateRandomOrdering` was not generating a
permutation of 1..n but rather this with a leading 0. E.g. {0, 3, 2, 4, 1}. Note
how the second loop in `generateRandomOrdering` was not touching the first
element of the array. This had no functional effect as this first element was
later ignored, but did involve an allocation of one more array entry than
needed.

This change removes the unused allocation and entry. However, 1-based indexing
is used pervasively in this segment code and should probably be more thoroughly
removed in future.