gvgen: fix incorrect bounds applied during tree generation
`genTree` contains the following line:

  double Z = floor(v * drand());

I do not fully understand the algorithm being implemented here, but it seems
clear the intent here is to generate a number in the range `[0, v)`. However
this was instead generating a number in the range `[0, v]`. That is, because
`rand()` can produce a value up to the _inclusive_ bound `RAND_MAX`, it was
possible for `drand()` to return `1.0`. This appears to have been the cause of
crashes when generating certain trees:

  $ gvgen -R 20 -u1967
  ==4236==ERROR: AddressSanitizer: heap-buffer-overflow on address 0x02c01af4 at pc 0x00c551e4 bp 0x006ff808 sp 0x006ff7fc
  READ of size 4 at 0x02c01af4 thread T0
    #0 0xc551e3 in genTree C:\GitLab-Runner\builds\smattr\graphviz\cmd\tools\graph_generator.c:608
    #1 0xc5277b in makeRandomTree C:\GitLab-Runner\builds\smattr\graphviz\cmd\tools\graph_generator.c:679
  …

Debugging the above reveals a sequence of assignments:

  v = 4.000000
  Z = 4.000000

indicating that `floor(v * drand())` indeed returned `v`.

Surprisingly this seems unrelated to the alteration of this code to use unsigned
values (2b61b45fb141bc019b1a75d15a3ed78b5fc7b56f) which would be the obvious
suspect. It seems this bug has existed since the very first introduction of this
code (da90960550147b67f70e9124db5306e88191b35a) but required some coincidences
to hit it and some instrumentation like ASan to even realize what went wrong.

Note that there are a number of other problems with the random number generation
occurring here that still remain:
  1. Floating-point precision means it is not necessarily possible for
     `floor(v * drand())` to produce all possible values in the range `[0, v)`.
  2. `(double)rand() / (RAND_MAX + 1)` does not result in a uniform distribution
     of random numbers across its range.
Both of these are left to be addressed in future changes.

This fixes `gvgen` runs with `-u1967`, but it also changes the sequence of
`drand()` results that would be generated with this seed. This change was
validated by running the Windows CI jobs over as much of the range (0, 2³¹) as
they could cover before timing out. It is possible future changes will introduce
problems with a specific seed that differs from 1967, so the #2640 test case is
updated to cover the space [1, 2000) that is assumed to be large enough to be
likely to find problematic seeds.

Gitlab: fixes #2640