util gv_random: fix handling of bounds > 'RAND_MAX'
`RAND_MAX` is only guaranteed to be at least 32767. So attempting to generate a
random number in a range that exceeded this could result in high values never
being generated. This is not hypothetical. One of the two existing `gv_random`
calls passes 100000.

This change allows `gv_random` to support the entire range of `int`.