CI: fix missing tcl-devel, tk-devel dependencies on Fedora
Since 2025-03-10, the Fedora CI jobs have begun failing because these packages
are no longer found. It seems they were previously being transitively installed
as a dependency of something else whose dependencies have recently changed.
While this difference is visible in the CI logs, it is not possible to
conclusively determine which package changes were responsible for this without
going package by package through the dependencies, as everything in CI is
installed in a monolithic `dnf` call.

Gitlab: fixes #2657
Suggested-by: Mark Hansen <mark@markhansen.co.nz>