CI: explicitly fail Windows CI job if subcommands fail
When running commands within a PowerShell script (or in this case, a Gitlab CI
recipe) there is a distinction in how the command being run is handled based on
whether it is a (1) cmdlet, (2) PowerShell script, or (3) external program
(EXE). `$ErrorActionPreference = "Stop"` is documented to only affect cmdlets,¹
but empirically it seems to have the effect of making external program failures
go unnoticed.²

This change tries to regain some sanity through universal error handling:
  1. The existing `$ErrorActionPreference = "Stop"` should make failing cmdlets
     fail the job.
  2. The existing `$PSDefaultParameterValues['*:ErrorAction']='Stop'` should
     make failing PowerShell scripts fail the job.
  3. Appending `if (-not $?) { throw "command failed" }` to each external
     program call in this change should make failing EXEs fail the job.

Repeating this verbatim step after every EXE run is verbose, but a future change
will aim to move most of this logic out of PowerShell, avoiding the above
discussed issues as much as possible.

Further reading:
  • https://stackoverflow.com/questions/9948517/how-to-stop-a-powershell-script-on-the-first-error
  • https://rkeithhill.wordpress.com/2009/08/03/effective-powershell-item-16-dealing-with-errors/
  • https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_parameters_default_values?view=powershell-7.5

¹ https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_preference_variables?view=powershell-7.5#erroractionpreference
² https://gitlab.com/graphviz/graphviz/-/issues/2656#note_2387322897

Gitlab: #2656