flip polarity of '-Tsvg_inline' external reference test case and xfail
It has been concluded that tying `-Tsvg_inline` and “should references to
external images be inlined?” together is incorrect and infers something the user
may not have intended and may find surprising. In future, the option to inline
references to external images may be re-implemented as an orthogonal feature.

Gitlab: #2636, #2647