Revert "core plugin: inline referenced images when rendering `-Tsvg_inline`"
This reverts commit 4f737e9118e13a1b9b7ab1e4c434fbf60757d392.

It has been concluded that tying `-Tsvg_inline` and “should references to
external images be inlined?” together is incorrect and infers something the user
may not have intended and may find surprising. In future, the option to inline
references to external images may be re-implemented as an orthogonal feature.

Gitlab: #2636, closes #2647