avoid dumping stdout when 'dot' test function fails
7b32fdda6b8993566bdc7d26d00fe18ac6b1a122 refactored this code to use the common
function `run_raw` for handling the executed subprocess. An unfortunate side
effect of this was that, if the subprocess (`dot`) returned non-zero, (possibly
binary) data written to stdout would be copied to the terminal. This causes two
primary problems:
  • Locally, this corrupts a user’s terminal requiring a `reset` afterwards to
    regain usability.
  • In CI, this irrelevant data exceeds Gitlab’s log limits causing truncation
    and loss of information.

This change reverts to something similar to the prior behavior.

Gitlab: closes #2658