replace 'run_raw' with 'subprocess.run' in a 'newrank' test
An unintended side effect of 7b32fdda6b8993566bdc7d26d00fe18ac6b1a122 was that
if this test ever fails it sprays binary data into stdout. Locally, this
corrupts the containing terminal. In CI, this exceeds log limits, truncating
output.

This change causes the PNG output to never be dumped to stdout, even in the case
of failure.

Gitlab: #2521, #2658