CI: port Windows utility dependency setup to Python
This is part of a series of changes to move Windows CI logic into Python. The
motivation for this is avoiding unexpected PowerShell behaviors that result in
unnoticed CI failures.

Note that ci/windows_build.py no longer needs to open the logfile in mode `"at"`
because now nothing writes to this file before this script’s execution. But this
is left as is because this code will be removed in an upcoming commit.

Gitlab: https://gitlab.com/graphviz/graphviz/-/issues/2656#note_2387358792