CI: port Windows test suite steps to Python
This is part of a series of changes to move Windows CI logic into Python. The
motivation for this is avoiding unexpected PowerShell behaviors that result in
unnoticed CI failures.

A surprising discovery during construction of this commit was that `$Path` needs
to be referenced as uppercase in Python on Windows (`os.environ["PATH"]`).
Other cased versions of this do not find pre-existing content
(`os.environ.get("Path") == None`). It is unclear to me exactly why this
happens, but the behavior seems to be known and intentional:
  • https://stackoverflow.com/questions/7797269/python-environment-case-senstivity-os-environ
  • https://stackoverflow.com/questions/19023238/why-python-uppercases-all-environment-variables-in-windows
  • https://bugs.python.org/issue28824

Gitlab: https://gitlab.com/graphviz/graphviz/-/issues/2656#note_2387358792