ABI break: update cdt’s SO name 5 → 6
Between the last release (12.2.1) and now, ABI breaks have occurred. E.g.
5b789236e871fec3fa82890b20f345abfdab218d. The ABI version of most Graphviz
libraries has not been updated in the past, even when it probably should have
been. This was implicitly left for downstream packagers to handle. This change
is an attempt to regain rigour with respect to how ABI breaks are communicated.