cgraph agdictsym: abbreviate 'key' initialization, zeroing other fields
This removes an implicit assumption that `dtsearch`-ing the dictionary here
would not read any other fields than `name` of the key. This is true currently
but left a sharp edge for anyone in future modifying the way this dictionary is
indexed. Naïve attempts to alter the fields used for search would have
introduced Undefined Behavior.