cgraph: restore 'agattr'
Discussion with the downstream project Pygraphviz has revealed that removing
`agattr` entirely is likely to be disruptive to Graphviz consumers in a way they
will not be sure how to resolve. A sketch of the scenario we are concerned about
is a downstream project that is calling `agattr` on behalf of some higher level
user request. Their wrapper around Graphviz’ `agattr` has no knowledge about
user intent. So the advice of “you now need to distinguish between regular text
attributes and HTML-like attributes” is not easy for them to apply when they do
not know which type of attribute any particular call from their user is trying
to create.

It was decided that reintroducing `agattr` as done within this change is a
reasonable compromise to smooth over this transition. For users who are not
creating regular text attributes and HTML-like attributes that alias one another
(the vast majority of users), this reintroduced `agattr` should behave the same
as the old one they were used to.

This replacement `agattr` has deliberately been defined below `agattr_text` and
`agattr_html` to encourage casual readers to encounter the more precise
alternatives first and default to using them.

Note that the `agmarkhtmlstr` function remains removed (done in
d1d4474cac299118730c6ea4dfefdb3096f6a106), so any consumers creating attributes
and then post hoc marking them as HTML-like will have a compilation/linking
break. We are not currently aware of any third-party use of this function. In
general, creating an attribute and then adjusting its properties this way was an
unwise pattern that was/is fundamentally incompatible with #2089.

Gitlab: #2089, #2641
Reported-by: Rob Hart <robhart@google.com>
Suggested-by: Rob Hart <robhart@google.com>