gvgen: avoid 'double' in tree generation; use uniform distribution for RNG
This code appears to have been written to use `double`, not because it needs
fractional values, but because it needs a wider range than `int`/`unsigned`.
With hindsight and progress in the C environment, we can rewrite this to use
64-bit integers. This is faster and neatly solves both of the remaining issues
mentioned in 8a5c4dd326060e77b2ac57920865456e83e2a6c4:
  1. It is now possible to generated all values in the range `[0, v)`.
  2. The generated random values fit a uniform distribution.

This change does not include a changelog entry, despite affecting the specific
sequence of random numbers that will be generated for any given `gvgen` run.
Prior to 5f210034c6c8c6a64121e353ee2856fa90999c75, it was not possible to
control seeding of the RNG in a deterministic way. So it is assumed that no
users of Graphviz ≤ 12.2.1 can be relying on the specific sequence of random
numbers used in tree generation.

Gitlab: #2640